[BUGFIX] Clear supply on market setup change
